Move material package generation to Celery
